pyproject.toml
[tool.poetry]
name = "polyus-test-task"
version = "0.1.0"
description = ""
authors = ["Алексей Лесков <alexleskov2000@gmail.com>"]
packages = [{ include = "polyus_nsi", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.31"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
alembic = "^1.13.2"
python-json-logger = "^2.0.7"
attrs = "^23.2.0"
uvicorn = "^0.30.1"
tzdata = "^2024.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
yapf = "^0.40.2"
flake8 = "^7.1.0"
toml = "^0.10.2"
flake8-pyproject = "^1.2.3"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
known_first_party = "polyus_nsi"
line_length = 80

[tool.yapf]
based_on_style = "pep8"
spaces_before_comment = 4
column_limit = 80
allow_split_before_dict_value = false
dedent_closing_brackets = true
split_complex_comprehension = true
split_arguments_when_comma_terminated = true
blank_line_before_nested_class_or_def = true
split_before_dict_set_generator = true
split_before_first_argument = true
split_before_arithmetic_operator = true
split_before_logical_operator = true
split_before_bitwise_operator = true

[tool.flake8]
max-line-length = 80
select = ['F', 'E', 'W', ]
ignore = ['W503', ]
per-file-ignores = [
    '*/adapters/*/repositories.py:E711,E712',
    '*/adapters/*/repositories/*:E711,E712',
    '*/migrations/*:E711,E712',
    '*/alembic/versions/*:E711,E712',
    '__init__.py:F401',
]
count = true